+++ views/landing.hbs
@@ -1,7 +1,6 @@
 <!DOCTYPE html>
 <html>
 
-
 <head>
     <title>Eative</title>
 
@@ -114,7 +113,7 @@
                 &nbsp;EATIVE&nbsp;
             </span>
         </h1>
-        <a href="/login" class="btn btn-lg btn-danger">View All Restaurants</a>
+        <a href="/restaurants" class="btn btn-lg btn-danger">View All Restaurants</a>
     </div>
 
     <ul class="slideshow">
