+++ views/restaurants-show.hbs
@@ -1,8 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
 
-
-
 <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
@@ -46,7 +44,7 @@
                         <a href="/restaurants">HOME</a>
                     </li>
                     <li>
-                        <a href= "#">ABOUT</a>
+                        <a href="https://www.linkedin.com/in/monika9/">ABOUT</a>
                     </li>
                     <li>
                         <a href="/owners/add">ADD RESTAURANT</a>
